third_party/blink/web_tests/fast/dom/SelectorAPI/loose-subtree-nth.html

<!DOCTYPE html>
<title>Element.querySelector matching nth-child of loose subtree.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
setup({ single_test: true });

var root = document.createElement("div");
for (var i = 0; i < 12; i++) {
  var child = document.createElement("span");
  if (i == 11)
    child.id = "pass";
  root.appendChild(child);
}

assert_equals(root.querySelector(":nth-child(12)").id, "pass");
done();
</script>
